ai-agent/requirements.txt
langchain>=1.2.0
langchain-groq>=1.1.0
langchain-community>=0.4.0
langchain-huggingface>=1.2.0
langchain-core>=1.2.0
chromadb>=1.5.0
sentence-transformers>=5.0.0
fastapi>=0.135.0
uvicorn>=0.42.0
requests>=2.33.0
pydantic>=2.12.0
python-dotenv>=1.2.0
